Struct object_store::chunked::ChunkedStore

pub struct ChunkedStore { /* private fields */ }

Wraps a ObjectStore and makes its get response return chunks in a controllable manner.

A ChunkedStore makes the memory consumption and performance of the wrapped ObjectStore worse. It is intended for use within tests, to control the chunks in the produced output streams. For example, it is used to verify the delimiting logic in newline_delimited_stream.

Implementations

impl ChunkedStore

pub fn new(inner: Arc<dyn ObjectStore>, chunk_size: usize) -> Self

Creates a new ChunkedStore with the specified chunk_size