object_store
Enum Metakey

pub enum Metakey {
    Complete,
    Mode,
    CacheControl,
    ContentDisposition,
    ContentLength,
    ContentMd5,
    ContentRange,
    ContentType,
    Etag,
    LastModified,
    Version,
    UserMetaData,
}

Metakey describes the metadata keys that can be stored or queried.

For store

Internally, we will store a flag set of Metakey to check whether we have set some key already.

For query

At user side, we will allow user to query the metadata. If the meta has been stored, we will return directly. If no, we will call stat internally to fetch the metadata.

Variants

Complete

The special metadata key that used to mark this entry already contains all metadata.

Mode

Key for mode.

CacheControl

Key for cache control.

ContentDisposition

Key for content disposition.

ContentLength

Key for content length.

ContentMd5

Key for content md5.

ContentRange

Key for content range.

ContentType

Key for content type.

Etag

Key for etag.

LastModified

Key for last modified.

Version

Key for version.

UserMetaData

Key for user metadata