Module object_store::layers

Modules

lru_cache 🔒

prometheus 🔒

code originally from https://github.com/apache/incubator-opendal/blob/main/core/src/layers/prometheus.rs, make a tiny change to avoid crash in multi thread env

Structs

ConcurrentLimitLayer

Add concurrent request limit.

ImmutableIndexLayer

Add an immutable in-memory index for underlying storage services.

LoggingLayer

Add log for every operation.

LruCacheAccess

LruCacheLayer

An opendal layer with local LRU file cache supporting.

PrometheusMetricsLayer

Please refer to prometheus for every operation.

RetryLayer

Add retry for temporary failed operations.

TimeoutLayer

Add timeout for every operation to avoid slow or unexpected hang operations.

TracingLayer

Add tracing for every operation.

Traits

LoggingInterceptor

LoggingInterceptor is used to intercept the log.

RetryInterceptor

RetryInterceptor is used to intercept while retry happened.