object_store::layers

Module mock

Modules

oio

oio provides OpenDAL’s raw traits and types that opendal returns as output.

Structs

AppendWriter

AppendWriter will implements oio::Write based on append object.

BatchDeleteResult

BatchDeleteResult is the result of batch delete operation.

BatchDeleter

BatchDeleter is used to implement oio::Delete based on batch delete.

BlockWriter

BlockWriter will implement oio::Write based on block uploads.

Buffer

Buffer is a wrapper of contiguous Bytes and non-contiguous [Bytes].

Entry

Entry is returned by Page or BlockingPage during list operations.

Error

Error is the error struct returned by all opendal functions.

FlatLister

FlatLister will walk dir in bottom up way:

FlexBuf

FlexBuf is a buffer that support frozen bytes and reuse existing allocated memory.

HierarchyLister

ToHierarchyLister will convert a flat list to hierarchy by filter not needed entries.

Metadata

Metadata contains all the information related to a specific path.

MockAccessor

MockDeleter

MockLayer

MockLayerBuilder

Builder for MockLayer.

MockLister

MockReader

MockWriter

MultipartPart

The result of MultipartWrite::write_part.

MultipartWriter

MultipartWriter will implement oio::Write based on multipart uploads.

OneShotDeleter

OneShotDelete is used to implement oio::Delete based on one shot.

OneShotWriter

OneShotWrite is used to implement oio::Write based on one shot.

OpDelete

Args for delete operation.

OpList

Args for list operation.

OpRead

Args for read operation.

OpWrite

Args for write operation.

PageContext

PageContext is the context passing between PageList.

PageLister

PageLister implements oio::List based on PageList.

PooledBuf

PooledBuf is a buffer pool that designed for reusing already allocated bufs.

PositionWriter

PositionWriter will implement oio::Write based on position write.

PrefixLister

PrefixLister is used to filter entries by prefix.

QueueBuf

QueueBuf is a queue of Buffer.

RpDelete

Reply for delete operation

RpList

Reply for list operation.

RpRead

Reply for read operation.

RpWrite

Reply for write operation.

Enums

ErrorKind

ErrorKind is all kinds of Error of opendal.

MockLayerBuilderError

Error type for MockLayerBuilder

Traits

Access

Underlying trait of all backends for implementers.

AppendWrite

AppendWrite is used to implement oio::Write based on append object. By implementing AppendWrite, services don’t need to care about the details of buffering and uploading parts.

BatchDelete

BatchDelete is used to implement oio::Delete based on batch delete operation.

BlockWrite

BlockWrite is used to implement oio::Write based on block uploads. By implementing BlockWrite, services don’t need to care about the details of uploading blocks.

Delete

The Delete trait defines interfaces for performing deletion operations.

DeleteDyn

The dyn version of Delete

Layer

Layer is used to intercept the operations on the underlying storage.

LayeredAccess

LayeredAccess is layered accessor that forward all not implemented method to inner.

List

Page trait is used by [raw::Accessor] to implement list operation.

MultipartWrite

MultipartWrite is used to implement oio::Write based on multipart uploads. By implementing MultipartWrite, services don’t need to care about the details of uploading parts.

OneShotDelete

OneShotDelete is used to implement oio::Delete based on one shot operation.

OneShotWrite

OneShotWrite is used to implement oio::Write based on one shot operation. By implementing OneShotWrite, services don’t need to care about the details.

PageList

PageList is used to implement oio::List based on API supporting pagination. By implementing PageList, services don’t need to care about the details of page list.

PositionWrite

PositionWrite is used to implement oio::Write based on position write.

Read

Read is the internal trait used by OpenDAL to read data from storage.

ReadDyn

ReadDyn is the dyn version of Read make it possible to use as Box<dyn ReadDyn>.

Write

Write is the trait that OpenDAL returns to callers.

Type Aliases

CopyInterceptor

Deleter

Deleter is a type erased Delete

Lister

The boxed version of List

MockDeleterFactory

MockListerFactory

MockReaderFactory

MockWriterFactory

Reader

Reader is a type erased Read.

Result

Result that is a wrapper of Result<T, opendal::Error>

Writer

Writer is a type erased Write