object_store::layers::mock

Trait ReadDyn

pub trait ReadDyn:
    Unpin
    + Send
    + Sync {
    // Required methods
    fn read_dyn(
        &mut self,
    ) -> Pin<Box<dyn Future<Output = Result<Buffer, Error>> + Send + '_>>;
    fn read_all_dyn(
        &mut self,
    ) -> Pin<Box<dyn Future<Output = Result<Buffer, Error>> + Send + '_>>;
}

ReadDyn is the dyn version of Read make it possible to use as Box<dyn ReadDyn>.

Required Methods

fn read_dyn( &mut self, ) -> Pin<Box<dyn Future<Output = Result<Buffer, Error>> + Send + '_>>

The dyn version of Read::read.

This function returns a boxed future to make it object safe.

fn read_all_dyn( &mut self, ) -> Pin<Box<dyn Future<Output = Result<Buffer, Error>> + Send + '_>>

The dyn version of Read::read_all