object_store::layers

Struct PrometheusLayer

pub struct PrometheusLayer {
    interceptor: PrometheusInterceptor,
}

Add prometheus for every operation.

Prometheus Metrics

We provide several metrics, please see the documentation of observe module. For a more detailed explanation of these metrics and how they are used, please refer to the Prometheus documentation.

Examples

let registry = prometheus::default_registry();

let op = Operator::new(services::Memory::default())?
    .layer(
        PrometheusLayer::builder()
            .register(registry)
            .expect("register metrics successfully"),
    )
    .finish();
debug!("operator: {op:?}");

// Write data into object test.
op.write("test", "Hello, World!").await?;
// Read data from object.
let bs = op.read("test").await?;
info!("content: {}", String::from_utf8_lossy(&bs.to_bytes()));

// Get object metadata.
let meta = op.stat("test").await?;
info!("meta: {:?}", meta);

// Export prometheus metrics.
let mut buffer = Vec::<u8>::new();
let encoder = prometheus::TextEncoder::new();
encoder.encode(&prometheus::gather(), &mut buffer).unwrap();
println!("## Prometheus Metrics");
println!("{}", String::from_utf8(buffer.clone()).unwrap());

Ok(())

Fields
interceptor: PrometheusInterceptor
Implementations

impl PrometheusLayer

pub fn builder() -> PrometheusLayerBuilder

Create a PrometheusLayerBuilder to set the configuration of metrics.

Default Configuration

operation_duration_seconds_buckets: exponential_buckets(0.01, 2.0, 16)

operation_bytes_buckets: exponential_buckets(1.0, 2.0, 16)

path_label: 0

Example

// Pick a builder and configure it.
let builder = services::Memory::default();
let registry = prometheus::default_registry();

let duration_seconds_buckets = prometheus::exponential_buckets(0.01, 2.0, 16).unwrap();
let bytes_buckets = prometheus::exponential_buckets(1.0, 2.0, 16).unwrap();
let op = Operator::new(builder)?
    .layer(
        PrometheusLayer::builder()
            .operation_duration_seconds_buckets(duration_seconds_buckets)
            .operation_bytes_buckets(bytes_buckets)
            .path_label(0)
            .register(registry)
            .expect("register metrics successfully"),
    )
    .finish();
debug!("operator: {op:?}");

Ok(())