Struct object_store::path::Path

pub struct Path { /* private fields */ }

A parsed path representation that can be safely written to object storage

Path Safety

In theory object stores support any UTF-8 character sequence, however, certain character sequences cause compatibility problems with some applications and protocols. As such the naming guidelines for S3, GCS and Azure Blob Storage all recommend sticking to a limited character subset.

This presents libraries with two options for consistent path handling:

Allow constructing unsafe paths, allowing for both reading and writing of data to paths that may not be consistently understood or supported

Disallow constructing unsafe paths, ensuring data written can be consistently handled by all other systems, but preventing interaction with objects at unsafe paths

This library takes the second approach, in particular:

Paths are delimited by /

Paths do not start with a /

Empty path segments are discarded (e.g. // is treated as though it were /)

Relative path segments, i.e. . and .. are percent encoded

Unsafe characters are percent encoded, as described by RFC 1738

All paths are relative to the root of the object store

In order to provide these guarantees there are two ways to safely construct a Path

Encode

A string containing potentially illegal path segments can be encoded to a Path using Path::from or Path::from_iter.

assert_eq!(Path::from("foo/bar").as_ref(), "foo/bar");
assert_eq!(Path::from("foo//bar").as_ref(), "foo/bar");
assert_eq!(Path::from("foo/../bar").as_ref(), "foo/%2E%2E/bar");
assert_eq!(Path::from_iter(["foo", "foo/bar"]).as_ref(), "foo/foo%2Fbar");

Note: if provided with an already percent encoded string, this will encode it again

assert_eq!(Path::from("foo/foo%2Fbar").as_ref(), "foo/foo%252Fbar");

Parse

Alternatively a Path can be created from an existing string, returning an error if it is invalid. Unlike the encoding methods, this will permit valid percent encoded sequences.

assert_eq!(Path::parse("/foo/foo%2Fbar").unwrap().as_ref(), "foo/foo%2Fbar");
Path::parse("..").unwrap_err();
Path::parse("/foo//").unwrap_err();
Path::parse("๐").unwrap_err();

Implementations

impl Path

pub fn parse(path: impl AsRef<str>) -> Result<Self, Error>

Parse a string as a Path, returning a Error if invalid, as defined on the docstring for Path

Note: this will strip any leading / or trailing /

pub fn from_filesystem_path(path: impl AsRef<Path>) -> Result<Self, Error>

Convert a filesystem path to a Path relative to the filesystem root

This will return an error if the path contains illegal character sequences as defined on the docstring for Path or does not exist

Note: this will canonicalize the provided path, resolving any symlinks

pub fn from_absolute_path(path: impl AsRef<Path>) -> Result<Self, Error>

Convert an absolute filesystem path to a Path relative to the filesystem root

This will return an error if the path contains illegal character sequences, as defined on the docstring for Path, or base is not an absolute path

pub fn from_url_path(path: impl AsRef<str>) -> Result<Self, Error>

Parse a url encoded string as a Path, returning a Error if invalid

This will return an error if the path contains illegal character sequences as defined on the docstring for Path

pub fn parts(&self) -> impl Iterator<Item = PathPart<'_>>

Returns the PathPart of this Path

pub fn filename(&self) -> Option<&str>

Returns the last path segment containing the filename stored in this Path

pub fn extension(&self) -> Option<&str>

Returns the extension of the file stored in this Path, if any

pub fn prefix_match( &self, prefix: &Self ) -> Option<impl Iterator<Item = PathPart<'_>> + '_>

Returns an iterator of the PathPart of this Path after prefix

Returns None if the prefix does not match

pub fn prefix_matches(&self, prefix: &Self) -> bool

Returns true if this Path starts with prefix

pub fn child<'a>(&self, child: impl Into<PathPart<'a>>) -> Self

Creates a new child of this Path