object_store::services

Struct Oss

pub struct Oss {
    config: OssConfig,
    http_client: Option<HttpClient>,
}

Aliyun Object Storage Service (OSS) support

Capabilities

This service can be used to:

stat

read

write

append

create_dir

delete

copy

rename

list

presign

blocking

Configuration

root: Set the work dir for backend.

bucket: Set the container name for backend.

endpoint: Set the endpoint for backend.

presign_endpoint: Set the endpoint for presign.

access_key_id: Set the access_key_id for backend.

access_key_secret: Set the access_key_secret for backend.

role_arn: Set the role of backend.

oidc_token: Set the oidc_token for backend.

allow_anonymous: Set the backend access OSS in anonymous way.

Refer to OssBuilderโs public API docs for more information.

Example

Via Builder

use std::sync::Arc;

use anyhow::Result;
use opendal::services::Oss;
use opendal::Operator;

#[tokio::main]
async fn main() -> Result<()> {
    // Create OSS backend builder.
    let mut builder = Oss::default()
        // Set the root for oss, all operations will happen under this root.
        //
        // NOTE: the root must be absolute path.
        .root("/path/to/dir")
        // Set the bucket name, this is required.
        .bucket("test")
        // Set the endpoint.
        //
        // For example:
        // - "https://oss-ap-northeast-1.aliyuncs.com"
        // - "https://oss-hangzhou.aliyuncs.com"
        .endpoint("https://oss-cn-beijing.aliyuncs.com")
        // Set the access_key_id and access_key_secret.
        //
        // OpenDAL will try load credential from the env.
        // If credential not set and no valid credential in env, OpenDAL will
        // send request without signing like anonymous user.
        .access_key_id("access_key_id")
        .access_key_secret("access_key_secret");

    let op: Operator = Operator::new(builder)?.finish();

    Ok(())
}

Fields
config: OssConfighttp_client: Option<HttpClient>
Implementations

impl OssBuilder

pub fn root(self, root: &str) -> OssBuilder

Set root of this backend.

All operations will happen under this root.

pub fn bucket(self, bucket: &str) -> OssBuilder

Set bucket name of this backend.

pub fn endpoint(self, endpoint: &str) -> OssBuilder

Set endpoint of this backend.

pub fn enable_versioning(self, enabled: bool) -> OssBuilder

Set bucket versioning status for this backend

pub fn presign_endpoint(self, endpoint: &str) -> OssBuilder

Set an endpoint for generating presigned urls.

You can offer a public endpoint like https://oss-cn-beijing.aliyuncs.com to return a presinged url for public accessors, along with an internal endpoint like https://oss-cn-beijing-internal.aliyuncs.com to access objects in a faster path.

If presign_endpoint is set, we will use presign_endpoint on generating presigned urls.

if not, we will use endpoint as default.

pub fn access_key_id(self, v: &str) -> OssBuilder

Set access_key_id of this backend.

If access_key_id is set, we will take userโs input first.

If not, we will try to load it from environment.

pub fn access_key_secret(self, v: &str) -> OssBuilder

Set access_key_secret of this backend.

If access_key_secret is set, we will take userโs input first.

If not, we will try to load it from environment.

pub fn http_client(self, client: HttpClient) -> OssBuilder

Specify the http client that used by this service.

Notes

This API is part of OpenDALโs Raw API. HttpClient could be changed during minor updates.

pub fn server_side_encryption(self, v: &str) -> OssBuilder

Set server_side_encryption for this backend.

Available values: AES256, KMS.

Reference: https://www.alibabacloud.com/help/en/object-storage-service/latest/server-side-encryption-5 Brief explanation: There are two server-side encryption methods available: SSE-AES256: 1. Configure the bucket encryption mode as OSS-managed and specify the encryption algorithm as AES256. 2. Include the x-oss-server-side-encryption parameter in the request and set its value to AES256. SSE-KMS: 1. To use this service, you need to first enable KMS. 2. Configure the bucket encryption mode as KMS, and specify the specific CMK ID for BYOK (Bring Your Own Key) or not specify the specific CMK ID for OSS-managed KMS key. 3. Include the x-oss-server-side-encryption parameter in the request and set its value to KMS. 4. If a specific CMK ID is specified, include the x-oss-server-side-encryption-key-id parameter in the request, and set its value to the specified CMK ID.

pub fn server_side_encryption_key_id(self, v: &str) -> OssBuilder

Set server_side_encryption_key_id for this backend.

Notes

This option only takes effect when server_side_encryption equals to KMS.

pub fn batch_max_operations(self, delete_max_size: usize) -> OssBuilder

๐Deprecated since 0.52.0: Please use delete_max_size instead of batch_max_operations

Set maximum batch operations of this backend.

pub fn delete_max_size(self, delete_max_size: usize) -> OssBuilder

Set maximum delete operations of this backend.

pub fn allow_anonymous(self) -> OssBuilder

Allow anonymous will allow opendal to send request without signing when credential is not loaded.

pub fn role_arn(self, role_arn: &str) -> OssBuilder

Set role_arn for this backend.

If role_arn is set, we will use already known config as source credential to assume role with role_arn.

pub fn role_session_name(self, role_session_name: &str) -> OssBuilder

Set role_session_name for this backend.

pub fn oidc_provider_arn(self, oidc_provider_arn: &str) -> OssBuilder

Set oidc_provider_arn for this backend.

pub fn oidc_token_file(self, oidc_token_file: &str) -> OssBuilder

Set oidc_token_file for this backend.

pub fn sts_endpoint(self, sts_endpoint: &str) -> OssBuilder

Set sts_endpoint for this backend.