Struct object_store::services::Webhdfs

pub struct Webhdfs {
    root: Option<String>,
    endpoint: Option<String>,
    delegation: Option<String>,
    disable_list_batch: bool,
}

WebHDFS’s REST API support. There two implementations of WebHDFS REST API:

Native via HDFS Namenode and Datanode, data are transferred between nodes directly.

HttpFS is a gateway before hdfs nodes, data are proxied.

Capabilities

This service can be used to:

stat

read

write

create_dir

delete

copy

rename

list

scan

presign

blocking

Differences with HDFS

[Hdfs][crate::services::Hdfs] is powered by HDFS’s native java client. Users need to set up the HDFS services correctly. But webhdfs can access from HTTP API and no extra setup needed.

Configurations

root: The root path of the WebHDFS service.

endpoint: The endpoint of the WebHDFS service.

delegation: The delegation token for WebHDFS.

Refer to Builder’s public API docs for more information.

Examples

Via Builder

use std::sync::Arc;

use anyhow::Result;
use opendal::services::Webhdfs;
use opendal::Operator;

#[tokio::main]
async fn main() -> Result<()> {
    let mut builder = Webhdfs::default();
    // set the root for WebHDFS, all operations will happen under this root
    //
    // Note:
    // if the root is not exists, the builder will automatically create the
    // root directory for you
    // if the root exists and is a directory, the builder will continue working
    // if the root exists and is a folder, the builder will fail on building backend
    builder.root("/path/to/dir");
    // set the endpoint of webhdfs namenode, controlled by dfs.namenode.http-address
    // default is http://127.0.0.1:9870
    builder.endpoint("http://127.0.0.1:9870");
    // set the delegation_token for builder
    builder.delegation("delegation_token");

    let op: Operator = Operator::new(builder)?.finish();

    Ok(())
}

Fields
root: Option<String>endpoint: Option<String>delegation: Option<String>disable_list_batch: bool
Implementations

impl WebhdfsBuilder

pub fn root(&mut self, root: &str) -> &mut WebhdfsBuilder

Set the working directory of this backend

All operations will happen under this root

Note

The root will be automatically created if not exists.

pub fn endpoint(&mut self, endpoint: &str) -> &mut WebhdfsBuilder

Set the remote address of this backend default to http://127.0.0.1:9870

Endpoints should be full uri, e.g.

https://webhdfs.example.com:9870

http://192.168.66.88:9870

If user inputs endpoint without scheme, we will prepend http:// to it.

pub fn delegation(&mut self, delegation: &str) -> &mut WebhdfsBuilder

Set the delegation token of this backend, used for authentication

Note

The builder prefers using delegation token over username. If both are set, delegation token will be used.

pub fn disable_list_batch(&mut self) -> &mut WebhdfsBuilder

Disable batch listing

Note

When listing a directory, the backend will default to use batch listing. If disable, the backend will list all files/directories in one request.