object_store

Struct Buffer

pub struct Buffer(Inner);

Buffer is a wrapper of contiguous Bytes and non-contiguous [Bytes].

We designed buffer to allow underlying storage to return non-contiguous bytes. For example, http based storage like s3 could generate non-contiguous bytes by stream.

Features

Buffer can be used as [Buf], Iterator, [Stream] directly.

Buffer is cheap to clone like [Bytes], only update reference count, no allocation.

Buffer is vectorized write friendly, you can convert it to IoSlice for vectored write.

Examples

As Buf

Buffer implements Buf trait:

use bytes::Buf;
use opendal::Buffer;
use serde_json;

fn test(mut buf: Buffer) -> Vec<String> {
    serde_json::from_reader(buf.reader()).unwrap()
}

As Bytes Iterator

Buffer implements Iterator<Item=Bytes> trait:

use bytes::Bytes;
use opendal::Buffer;

fn test(mut buf: Buffer) -> Vec<Bytes> {
    buf.into_iter().collect()
}

As Bytes Stream

Buffer implements Stream<Item=Result<Bytes, Infallible>> trait:

use bytes::Bytes;
use futures::TryStreamExt;
use opendal::Buffer;

async fn test(mut buf: Buffer) -> Vec<Bytes> {
    buf.into_iter().try_collect().await.unwrap()
}

As one contiguous Bytes

Buffer can make contiguous by transform into Bytes or Vec<u8>. Please keep in mind that this operation involves new allocation and bytes copy, and we can’t reuse the same memory region anymore.

use bytes::Bytes;
use opendal::Buffer;

fn test_to_vec(buf: Buffer) -> Vec<u8> {
    buf.to_vec()
}

fn test_to_bytes(buf: Buffer) -> Bytes {
    buf.to_bytes()
}

Tuple Fields
0: Inner
Implementations

impl Buffer

pub const fn new() -> Buffer

Create a new empty buffer.

This operation is const and no allocation will be performed.

pub fn len(&self) -> usize

Get the length of the buffer.

pub fn is_empty(&self) -> bool

Check if buffer is empty.

pub fn count(&self) -> usize

Number of [Bytes] in Buffer.

For contiguous buffer, it’s always 1. For non-contiguous buffer, it’s number of bytes available for use.

pub fn current(&self) -> Bytes

Get current [Bytes].

pub fn truncate(&mut self, len: usize)

Shortens the buffer, keeping the first len bytes and dropping the rest.

If len is greater than the buffer’s current length, this has no effect.

pub fn slice(&self, range: impl RangeBounds<usize>) -> Buffer

Returns a slice of self for the provided range.

This will increment the reference count for the underlying memory and return a new Buffer handle set to the slice.

This operation is O(1).

pub fn to_bytes(&self) -> Bytes

Combine all bytes together into one single [Bytes].

This operation is zero copy if the underlying bytes are contiguous. Otherwise, it will copy all bytes into one single [Bytes]. Please use API from [Buf], Iterator or [Stream] whenever possible.

pub fn to_vec(&self) -> Vec<u8>

Combine all bytes together into one single Vec<u8>.

This operation is not zero copy, it will copy all bytes into one single Vec<u8>. Please use API from [Buf], Iterator or [Stream] whenever possible.

pub fn to_io_slice(&self) -> Vec<IoSlice<'_>>

Convert buffer into a slice of IoSlice for vectored write.