object_store

Struct Error

pub struct Error {
    kind: ErrorKind,
    message: String,
    status: ErrorStatus,
    operation: &'static str,
    context: Vec<(&'static str, String)>,
    source: Option<Error>,
    backtrace: Backtrace,
}

Error is the error struct returned by all opendal functions.

Display

Error can be displayed in two ways:

Via Display: like err.to_string() or format!("{err}")

Error will be printed in a single line:

Unexpected, context: { path: /path/to/file, called: send_async } => something wrong happened, source: networking error"

Via Debug: like format!("{err:?}")

Error will be printed in multi lines with more details and backtraces (if captured):

Unexpected => something wrong happened

Context:
   path: /path/to/file
   called: send_async

Source: networking error

Backtrace:
   0: opendal::error::Error::new
             at ./src/error.rs:197:24
   1: opendal::error::tests::generate_error
             at ./src/error.rs:241:9
   2: opendal::error::tests::test_error_debug_with_backtrace::{{closure}}
             at ./src/error.rs:305:41
   ...

For conventional struct-style Debug representation, like format!("{err:#?}"):

Error {
    kind: Unexpected,
    message: "something wrong happened",
    status: Permanent,
    operation: "Read",
    context: [
        (
            "path",
            "/path/to/file",
        ),
        (
            "called",
            "send_async",
        ),
    ],
    source: Some(
        "networking error",
    ),
}

Fields
kind: ErrorKindmessage: Stringstatus: ErrorStatusoperation: &'static strcontext: Vec<(&'static str, String)>source: Option<Error>backtrace: Backtrace
Implementations

impl Error

pub fn new(kind: ErrorKind, message: impl Into<String>) -> Error

Create a new Error with error kind and message.

pub fn with_operation(self, operation: impl Into<&'static str>) -> Error

Update errorโs operation.

Notes

If the error already carries an operation, we will push a new context (called, operation).

pub fn with_context(self, key: &'static str, value: impl ToString) -> Error

Add more context in error.

pub fn set_source(self, src: impl Into<Error>) -> Error

Set source for error.

Notes

If the source has been set, we will raise a panic here.

pub fn map<F>(self, f: F) -> Error
where F: FnOnce(Error) -> Error,

Operate on error with map.

pub fn set_permanent(self) -> Error

Set permanent status for error.

pub fn set_temporary(self) -> Error

Set temporary status for error.

By set temporary, we indicate this error is retryable.

pub fn set_persistent(self) -> Error

Set persistent status for error.

By setting persistent, we indicate the retry should be stopped.

pub fn kind(&self) -> ErrorKind

Return errorโs kind.

pub fn is_temporary(&self) -> bool

Check if this error is temporary.