object_store

Struct FuturesAsyncReader

pub struct FuturesAsyncReader {
    ctx: Arc<ReadContext>,
    stream: BufferStream,
    buf: Buffer,
    start: u64,
    end: u64,
    pos: u64,
}

FuturesAsyncReader is the adapter of [AsyncRead], [AsyncBufRead] and [AsyncSeek] generated by Reader::into_futures_async_read.

Users can use this adapter in cases where they need to use [AsyncRead] related trait. FuturesAsyncReader reuses the same concurrent and chunk settings from Reader.

FuturesAsyncReader also implements Unpin, Send and Sync

Fields
ctx: Arc<ReadContext>stream: BufferStreambuf: Bufferstart: u64end: u64pos: u64