Struct object_store::ObjectMetadata

pub struct ObjectMetadata {
    complete: bool,
    mode: ObjectMode,
    content_length: Option<u64>,
    content_md5: Option<String>,
    content_type: Option<String>,
    content_range: Option<BytesContentRange>,
    last_modified: Option<OffsetDateTime>,
    etag: Option<String>,
    content_disposition: Option<String>,
}

Metadata carries all object metadata.

Notes

mode and content_length are required metadata that all services should provide during stat operation. But in list operation, a.k.a., Entry’s content length could be None.

Fields
complete: boolmode: ObjectModecontent_length: Option<u64>content_md5: Option<String>content_type: Option<String>content_range: Option<BytesContentRange>last_modified: Option<OffsetDateTime>etag: Option<String>content_disposition: Option<String>
Implementations

impl ObjectMetadata

pub fn new(mode: ObjectMode) -> ObjectMetadata

Create a new object metadata

pub fn is_complete(&self) -> bool

If this object metadata if complete

pub fn set_complete(&mut self) -> &mut ObjectMetadata

Make this object metadata if complete.

pub fn with_complete(self) -> ObjectMetadata

Make this object metadata if complete.

pub fn mode(&self) -> ObjectMode

Object mode represent this object’s mode.

pub fn set_mode(&mut self, mode: ObjectMode) -> &mut ObjectMetadata

Set mode for object.

pub fn with_mode(self, mode: ObjectMode) -> ObjectMetadata

Set mode for object.

pub fn content_length(&self) -> u64

Content length of this object.

Content-Length is defined by RFC 7230 Refer to MDN Content-Length for more information.

pub fn set_content_length(&mut self, content_length: u64) -> &mut ObjectMetadata

Set content length of this object.

pub fn with_content_length(self, content_length: u64) -> ObjectMetadata

Set content length of this object.

pub fn content_md5(&self) -> Option<&str>

Content MD5 of this object.

Content MD5 is defined by RFC 2616. And removed by RFC 7231.

OpenDAL will try its best to set this value, but not guarantee this value is the md5 of content.

pub fn set_content_md5(&mut self, content_md5: &str) -> &mut ObjectMetadata

Set content MD5 of this object.

Content MD5 is defined by RFC 2616. And removed by RFC 7231.

pub fn with_content_md5(self, content_md5: &str) -> ObjectMetadata

Set content MD5 of this object.

Content MD5 is defined by RFC 2616. And removed by RFC 7231.

pub fn content_type(&self) -> Option<&str>

Content Type of this object.

Content Type is defined by RFC 9110.

pub fn set_content_type(&mut self, v: &str) -> &mut ObjectMetadata

Set Content Type of this object.

Content Type is defined by RFC 9110.

pub fn with_content_type(self, v: &str) -> ObjectMetadata

Set Content Type of this object.

Content Type is defined by RFC 9110.

pub fn content_range(&self) -> Option<BytesContentRange>

Content Range of this object.

Content Range is defined by RFC 9110.

pub fn set_content_range(&mut self, v: BytesContentRange) -> &mut ObjectMetadata

Set Content Range of this object.

Content Range is defined by RFC 9110.

pub fn with_content_range(self, v: BytesContentRange) -> ObjectMetadata

Set Content Range of this object.

Content Range is defined by RFC 9110.

pub fn last_modified(&self) -> Option<OffsetDateTime>

Last modified of this object.

Last-Modified is defined by RFC 7232 Refer to MDN Last-Modified for more information.

OpenDAL parse the raw value into [OffsetDateTime] for convenient.

pub fn set_last_modified( &mut self, last_modified: OffsetDateTime ) -> &mut ObjectMetadata

Set Last modified of this object.

Last-Modified is defined by RFC 7232 Refer to MDN Last-Modified for more information.

pub fn with_last_modified(self, last_modified: OffsetDateTime) -> ObjectMetadata

Set Last modified of this object.

Last-Modified is defined by RFC 7232 Refer to MDN Last-Modified for more information.

pub fn etag(&self) -> Option<&str>

ETag of this object.

ETag is defined by RFC 7232 Refer to MDN ETag for more information.

OpenDAL will return this value AS-IS like the following:

"33a64df551425fcc55e4d42a148795d9f25f89d4"

W/"0815"

" is part of etag.

pub fn set_etag(&mut self, etag: &str) -> &mut ObjectMetadata

Set ETag of this object.

ETag is defined by RFC 7232 Refer to MDN ETag for more information.

OpenDAL will return this value AS-IS like the following:

"33a64df551425fcc55e4d42a148795d9f25f89d4"

W/"0815"

" is part of etag, don’t trim it before setting.

pub fn with_etag(self, etag: &str) -> ObjectMetadata

Set ETag of this object.

ETag is defined by RFC 7232 Refer to MDN ETag for more information.

OpenDAL will return this value AS-IS like the following:

"33a64df551425fcc55e4d42a148795d9f25f89d4"

W/"0815"

" is part of etag, don’t trim it before setting.

pub fn content_disposition(&self) -> Option<&str>

Content-Disposition of this object

Content-Disposition is defined by RFC 2616 and clarified usage in RFC 6266. Refer to MDN Content-Disposition for more information.

OpenDAL will return this value AS-IS like the following:

“inline”

“attachment”

“attachment; filename="filename.jpg"”

pub fn with_content_disposition( self, content_disposition: &str ) -> ObjectMetadata

Set Content-Disposition of this object

Content-Disposition is defined by RFC 2616 and clarified usage in RFC 6266. Refer to MDN Content-Disposition for more information.

OpenDAL will return this value AS-IS like the following:

“inline”

“attachment”

“attachment; filename="filename.jpg"”

pub fn set_content_disposition( &mut self, content_disposition: &str ) -> &mut ObjectMetadata

Set Content-Disposition of this object

Content-Disposition is defined by RFC 2616 and clarified usage in RFC 6266. Refer to MDN Content-Disposition for more information.

OpenDAL will return this value AS-IS like the following:

“inline”

“attachment”

“attachment; filename="filename.jpg"”