Struct object_store::Writer

pub struct Writer {
    inner: Box<dyn Write>,
}

Writer is designed to write data into given path in an asynchronous manner.

Notes

Please make sure either close or abort has been called before dropping the writer otherwise the data could be lost.

Usage

Write Multiple Chunks

Some services support to write multiple chunks of data into given path. Services that doesn’t support write multiple chunks will return ErrorKind::Unsupported error when calling write at the second time.

let mut w = op.writer("path/to/file").await?;
w.write(bs).await?;
w.write(bs).await?;
w.close().await?

Our writer also provides Writer::sink and Writer::copy support.

Besides, our writer implements [AsyncWrite] and [tokio::io::AsyncWrite].

Write with append enabled

Writer also supports to write with append enabled. This is useful when users want to append some data to the end of the file.

If file doesn’t exist, it will be created and just like calling write.

If file exists, data will be appended to the end of the file.

Possible Errors:

Some services store normal file and appendable file in different way. Trying to append on non-appendable file could return ErrorKind::ConditionNotMatch error.

Services that doesn’t support append will return ErrorKind::Unsupported error when creating writer with append enabled.

Fields
inner: Box<dyn Write>
Implementations

impl Writer

pub async fn write(&mut self, bs: impl Into<Bytes>) -> Result<(), Error>

Write into inner writer.

pub async fn sink<S, T>(&mut self, sink_from: S) -> Result<u64, Error>
where S: Stream<Item = Result<T, Error>>, T: Into<Bytes>,

Sink into writer.

sink will read data from given streamer and write them into writer directly without extra in-memory buffer.

Notes

Sink doesn’t support to be used with write concurrently.

Sink doesn’t support to be used without content length now.

Examples

use std::io::Result;

use bytes::Bytes;
use futures::stream;
use futures::StreamExt;
use opendal::Operator;

#[tokio::main]
async fn sink_example(op: Operator) -> Result<()> {
    let mut w = op.writer_with("path/to/file").await?;
    let stream = stream::iter(vec![vec![0; 4096], vec![1; 4096]]).map(Ok);
    w.sink(stream).await?;
    w.close().await?;
    Ok(())
}

pub async fn copy<R>(&mut self, read_from: R) -> Result<u64, Error>
where R: AsyncRead,

Copy into writer.

copy will read data from given reader and write them into writer directly with only one constant in-memory buffer.

Notes

Copy doesn’t support to be used with write concurrently.

Copy doesn’t support to be used without content length now.

Examples

use std::io::Result;

use bytes::Bytes;
use futures::io::Cursor;
use futures::stream;
use futures::StreamExt;
use opendal::Operator;

#[tokio::main]
async fn copy_example(op: Operator) -> Result<()> {
    let mut w = op.writer_with("path/to/file").await?;
    let reader = Cursor::new(vec![0; 4096]);
    w.copy(reader).await?;
    w.close().await?;
    Ok(())
}

pub async fn abort(&mut self) -> Result<(), Error>

Abort the writer and clean up all written data.

Notes

Abort should only be called when the writer is not closed or aborted, otherwise an unexpected error could be returned.

pub async fn close(&mut self) -> Result<(), Error>

Close the writer and make sure all data have been committed.

Notes

Close should only be called when the writer is not closed or aborted, otherwise an unexpected error could be returned.