Crate object_store

Re-exports

pub use config::AzblobConnection;

pub use config::GcsConnection;

pub use config::OssConnection;

pub use config::S3Connection;

Modules

config

error

factory

layers

manager

metrics 🔒

object-store metrics

services

Services will provide builders to build underlying backends.

test_util

util

Structs

Buffer

Buffer is a wrapper of contiguous Bytes and non-contiguous [Bytes].

Entry

Entry returned by Lister or [BlockingLister] to represent a path and it’s relative metadata.

Error

Error is the error struct returned by all opendal functions.

FuturesAsyncReader

FuturesAsyncReader is the adapter of [AsyncRead], [AsyncBufRead] and [AsyncSeek] generated by Reader::into_futures_async_read.

FuturesAsyncWriter

FuturesIoAsyncWriter is the adapter of [AsyncWrite] for Writer.

HttpClient

A HTTP client instance for OpenDAL’s services.

Lister

Lister is designed to list entries at given path in an asynchronous manner.

ObjectStore

The Operator serves as the entry point for all public asynchronous APIs.

Reader

Reader is designed to read data from given path in an asynchronous manner.

Writer

Writer is designed to write data into given path in an asynchronous manner.

Enums

EntryMode

EntryMode represents the mode.

ErrorKind

ErrorKind is all kinds of Error of opendal.

Scheme

Services that OpenDAL supports

Constants

ATOMIC_WRITE_DIR

OBJECT_CACHE_DIR

The default object cache directory name.

OLD_ATOMIC_WRITE_DIR

For compatibility. Remove this after a major version release.

Traits

Access

Underlying trait of all backends for implementers.

ObjectStoreBuilder

Builder is used to set up underlying services.

Type Aliases

Result

Result that is a wrapper of Result<T, opendal::Error>