Crate object_store

Modules

layers

manager

metrics 🔒

object-store metrics

services

Services will provide builders to build underlying backends.

test_util

util

Structs

Entry

Entry returned by Lister or [BlockingLister] to represent a path and it’s relative metadata.

Error

Error is the error struct returned by all opendal functions.

FuturesAsyncReader

FuturesAsyncReader is the adapter of [AsyncRead], [AsyncBufRead] and [AsyncSeek] generated by Reader::into_futures_async_read.

FuturesAsyncWriter

FuturesIoAsyncWriter is the adapter of [AsyncWrite] for Writer.

HttpClient

HttpClient that used across opendal.

Lister

Lister is designed to list entries at given path in an asynchronous manner.

ObjectStore

Operator is the entry for all public async APIs.

Reader

Reader is designed to read data from given path in an asynchronous manner.

Writer

Writer is designed to write data into given path in an asynchronous manner.

Enums

EntryMode

EntryMode represents the mode.

ErrorKind

ErrorKind is all kinds of Error of opendal.

Metakey

Metakey describes the metadata keys that can be stored or queried.

Traits

ObjectStoreBuilder

Builder is used to set up a real underlying service, i.e. storage accessor.

Functions

raw_normalize_path

Make sure all operation are constructed by normalized path:

Type Aliases

Result

Result that is a wrapper of Result<T, opendal::Error>