Crate object_store

Modules

cache_policy

layers

Layer is the mechanism to intercept operations.

services

Services will provide builders to build underlying backends.

test_util

util

Structs

Error

Error is the error struct returned by all opendal functions.

Object

Object is the handler for all object related operations.

ObjectLister

ObjectLister is returned by Object::list to list objects.

ObjectMetadata

Metadata carries all object metadata.

ObjectStore

Operator is the user-facing APIs for object and object streams.

Enums

ErrorKind

ErrorKind is all kinds of opendal’s Error.

ObjectMode

ObjectMode represents the corresponding object’s mode.

Traits

ObjectStoreBuilder

Builder will build an accessor;

Type Definitions

Result

Result that is a wrapper of Reustl<T, opendal::Error>