Crate object_store

Modules

layers

metrics 🔒

object-store metrics

services

Services will provide builders to build underlying backends.

test_util

util

Structs

Entry

Entry is the file/dir entry returned by Lister.

Error

Error is the error struct returned by all opendal functions.

ObjectStore

Operator is the entry for all public async APIs.

Reader

Reader is designed to read data from given path in an asynchronous manner.

Writer

Writer is designed to write data into given path in an asynchronous manner.

Enums

EntryMode

EntryMode represents the mode.

ErrorKind

ErrorKind is all kinds of Error of opendal.

Metakey

Metakey describes the metadata keys that can be stored or queried.

Traits

ObjectStoreBuilder

Builder is used to set up a real underlying service, i.e. storage accessor.

Functions

raw_normalize_path

Make sure all operation are constructed by normalized path:

Type Definitions

Pager

The boxed version of [Page]

Result

Result that is a wrapper of Result<T, opendal::Error>