object_store

Module layers

Modules

mock

observe

OpenDAL Observability

prometheus 🔒

Structs

CapabilityCheckLayer

Add an extra capability check layer for every operation

ConcurrentLimitLayer

Add concurrent request limit.

HttpClientLayer

Layer for replacing the default HTTP client with a custom one.

ImmutableIndexLayer

Add an immutable in-memory index for underlying storage services.

LoggingLayer

Add log for every operation.

PrometheusLayer

Add prometheus for every operation.

PrometheusLayerBuilder

PrometheusLayerBuilder is a config builder to build a PrometheusLayer.

RetryLayer

Add retry for temporary failed operations.

TimeoutLayer

Add timeout for every operation to avoid slow or unexpected hang operations.

TracingLayer

Add tracing for every operation.

Traits

LoggingInterceptor

LoggingInterceptor is used to intercept the log.

RetryInterceptor

RetryInterceptor is used to intercept while retry happened.

Functions

build_prometheus_metrics_layer

This logical tries to extract parent path from the object storage operation the function also relies on assumption that the region path is built from pattern <data|index>/catalog/schema/table_id/... OR greptimedb/object_cache/<read|write>/...