object_store::layers::lru_cache

Module read_cache

Structs

CacheAwareDeleter

ReadCache 🔒

Local read cache for files in object storage

Enums

ReadResult 🔒

Cache value for read file

Constants

READ_CACHE_DIR 🔒

Subdirectory of cached files for read.

RECOVER_CACHE_LIST_CONCURRENT 🔒

Functions

can_cache 🔒

Returns true when the path of the file can be cached.

read_cache_key 🔒

Generate a unique cache key for the read path and range.

read_cache_key_prefix 🔒

read_cache_root 🔒

to_output_reader 🔒