object_store::layers

Module observe

OpenDAL Observability Layer

This module offers essential components to facilitate the implementation of observability in OpenDAL.

Prometheus Metrics

These metrics are essential for understanding the behavior and performance of our applications.

Metric Name	Type	Description	Labels

operation_duration_seconds	Histogram	Histogram of time spent during opendal operations	scheme, namespace, root, operation, path

operation_bytes.	Histogram	Histogram of the bytes transferred during opendal operations	scheme, operation, root, operation, path

operation_errors_total	Counter	Error counter during opendal operations	scheme, operation, root, operation, path, error

Structs

MetricMetadata

The metric metadata which contains the metric name and help.

MetricsAccessor

The metrics accessor for opendal.

MetricsLayer

The metrics layer for opendal.

Statics

LABEL_ERROR

The metric label for the error kind.

LABEL_NAMESPACE

The metric label for the namespace like bucket name in s3.

LABEL_OPERATION

The metric label for the operation like read, write, list.

LABEL_PATH

The metric label for the path used by request.

LABEL_ROOT

The metric label for the root path.

LABEL_SCHEME

The metric label for the scheme like s3, fs, cos.

METRIC_OPERATION_BYTES

The metric metadata for the operation bytes.

METRIC_OPERATION_DURATION_SECONDS

The metric metadata for the operation duration in seconds.

METRIC_OPERATION_ERRORS_TOTAL

The metric metadata for the operation errors total.

Traits

MetricsIntercept

The interceptor for metrics.

Functions

path_label_value

Return the path label value according to the given path and level.