Struct object_store::ObjectStore

pub struct ObjectStore {
    accessor: Arc<dyn Accessor<BlockingPager = Box<dyn BlockingPage>, BlockingWriter = Box<dyn BlockingWrite>, Writer = Box<dyn Write>, Pager = Box<dyn Page>, Reader = Box<dyn Read>, BlockingReader = Box<dyn BlockingRead>>>,
    limit: usize,
}

Operator is the entry for all public async APIs.

Developer should manipulate the data from storage service through Operator only by right.

We will usually do some general checks and data transformations in this layer, like normalizing path from input, checking whether the path refers to one file or one directory, and so on. Read [concepts][docs::concepts] for more about Operator.

Examples

Read more backend init examples in services

use opendal::services::Fs;
use opendal::Operator;
#[tokio::main]
async fn main() -> Result<()> {
    // Create fs backend builder.
    let mut builder = Fs::default();
    // Set the root for fs, all operations will happen under this root.
    //
    // NOTE: the root must be absolute path.
    builder.root("/tmp");

    // Build an `Operator` to start operating the storage.
    let _: Operator = Operator::new(builder)?.finish();

    Ok(())
}

Fields
accessor: Arc<dyn Accessor<BlockingPager = Box<dyn BlockingPage>, BlockingWriter = Box<dyn BlockingWrite>, Writer = Box<dyn Write>, Pager = Box<dyn Page>, Reader = Box<dyn Read>, BlockingReader = Box<dyn BlockingRead>>>limit: usize
Implementations

impl Operator

Operator basic API.

pub fn limit(&self) -> usize

Get current operator’s limit. Limit is usually the maximum size of data that operator will handle in one operation.

pub fn with_limit(&self, limit: usize) -> Operator

Specify the batch limit.

Default: 1000

pub fn info(&self) -> OperatorInfo

Get information of underlying accessor.

Examples

use opendal::Operator;

let info = op.info();

pub fn blocking(&self) -> BlockingOperator

Create a new blocking operator.

This operation is nearly no cost.

impl Operator

Operator async API.

pub async fn check(&self) -> Result<(), Error>

Check if this operator can work correctly.

We will send a list request to path and return any errors we met.

use opendal::Operator;

op.check().await?;

pub async fn stat(&self, path: &str) -> Result<Metadata, Error>

Get current path’s metadata.

Notes

For fetch metadata of entries returned by Lister, it’s better to use [list_with] and [lister_with] with metakey query like Metakey::ContentLength | Metakey::LastModified so that we can avoid extra requests.

Examples

use opendal::ErrorKind;
if let Err(e) = op.stat("test").await {
    if e.kind() == ErrorKind::NotFound {
        println!("file not exist")
    }
}

pub fn stat_with(&self, path: &str) -> FutureStat

Get current path’s metadata without cache directly with extra options.

Notes

Use stat if you:

Want to detect the outside changes of path.

Don’t want to read from cached metadata.

You may want to use metadata if you are working with entries returned by Lister. It’s highly possible that metadata you want has already been cached.

Examples

use opendal::ErrorKind;
if let Err(e) = op.stat_with("test").if_match("<etag>").await {
    if e.kind() == ErrorKind::NotFound {
        println!("file not exist")
    }
}

pub async fn is_exist(&self, path: &str) -> Result<bool, Error>

Check if this path exists or not.

Example

use anyhow::Result;
use futures::io;
use opendal::Operator;

#[tokio::main]
async fn test(op: Operator) -> Result<()> {
    let _ = op.is_exist("test").await?;

    Ok(())
}

pub async fn create_dir(&self, path: &str) -> Result<(), Error>

Create a dir at given path.

Notes

To indicate that a path is a directory, it is compulsory to include a trailing / in the path. Failure to do so may result in NotADirectory error being returned by OpenDAL.

Behavior

Create on existing dir will succeed.

Create dir is always recursive, works like mkdir -p

Examples

op.create_dir("path/to/dir/").await?;

pub async fn read(&self, path: &str) -> Result<Vec<u8>, Error>

Read the whole path into a bytes.

This function will allocate a new bytes internally. For more precise memory control or reading data lazily, please use Operator::reader

Examples

let bs = op.read("path/to/file").await?;

pub fn read_with(&self, path: &str) -> FutureRead

Read the whole path into a bytes with extra options.

This function will allocate a new bytes internally. For more precise memory control or reading data lazily, please use Operator::reader

Examples

let bs = op.read_with("path/to/file").await?;
let bs = op.read_with("path/to/file").range(0..10).await?;

pub async fn reader(&self, path: &str) -> Result<Reader, Error>

Create a new reader which can read the whole path.

Examples

let r = op.reader("path/to/file").await?;

pub fn reader_with(&self, path: &str) -> FutureReader

Create a new reader with extra options

Examples

let r = op.reader_with("path/to/file").range((0..10)).await?;

pub async fn write(&self, path: &str, bs: impl Into<Bytes>) -> Result<(), Error>

Write bytes into path.

Notes

Write will make sure all bytes has been written, or an error will be returned.

Examples

use bytes::Bytes;

op.write("path/to/file", vec![0; 4096]).await?;

pub async fn copy(&self, from: &str, to: &str) -> Result<(), Error>

Copy a file from from to to.

Notes

from and to must be a file.

to will be overwritten if it exists.

If from and to are the same, an IsSameFile error will occur.

copy is idempotent. For same from and to input, the result will be the same.

Examples

op.copy("path/to/file", "path/to/file2").await?;

pub async fn rename(&self, from: &str, to: &str) -> Result<(), Error>

Rename a file from from to to.

Notes

from and to must be a file.

to will be overwritten if it exists.

If from and to are the same, an IsSameFile error will occur.

Examples

op.rename("path/to/file", "path/to/file2").await?;

pub async fn writer(&self, path: &str) -> Result<Writer, Error>

Write multiple bytes into path.

Refer to Writer for more details.

Examples

use bytes::Bytes;

let mut w = op.writer("path/to/file").await?;
w.write(vec![0; 4096]).await?;
w.write(vec![1; 4096]).await?;
w.close().await?;

pub fn writer_with(&self, path: &str) -> FutureWriter

Write multiple bytes into path with extra options.

Refer to Writer for more details.

Examples

use bytes::Bytes;

let mut w = op
    .writer_with("path/to/file")
    .content_type("application/octet-stream")
    .await?;
w.write(vec![0; 4096]).await?;
w.write(vec![1; 4096]).await?;
w.close().await?;

pub fn write_with(&self, path: &str, bs: impl Into<Bytes>) -> FutureWrite

Write data with extra options.

Notes

Write will make sure all bytes has been written, or an error will be returned.

Examples

use bytes::Bytes;

let bs = b"hello, world!".to_vec();
let _ = op
    .write_with("path/to/file", bs)
    .content_type("text/plain")
    .await?;

pub async fn delete(&self, path: &str) -> Result<(), Error>

Delete the given path.

Notes

Deleting a file that does not exist won’t return errors.

Examples

op.delete("test").await?;

pub fn delete_with(&self, path: &str) -> FutureDelete

Delete the given path with extra options.

Notes

Deleting a file that does not exist won’t return errors.

Examples

op.delete_with("test").await?;

pub async fn remove(&self, paths: Vec<String>) -> Result<(), Error>

Notes

If underlying services support delete in batch, we will use batch delete instead.

Examples

op.remove(vec!["abc".to_string(), "def".to_string()])
    .await?;

pub async fn remove_via( &self, input: impl Stream<Item = String> + Unpin ) -> Result<(), Error>

remove will remove files via the given paths.

remove_via will remove files via the given stream.

We will delete by chunks with given batch limit on the stream.

Notes

If underlying services support delete in batch, we will use batch delete instead.

Examples

use futures::stream;
let stream = stream::iter(vec!["abc".to_string(), "def".to_string()]);
op.remove_via(stream).await?;

pub async fn remove_all(&self, path: &str) -> Result<(), Error>

Remove the path and all nested dirs and files recursively.

Notes

If underlying services support delete in batch, we will use batch delete instead.

Examples

op.remove_all("path/to/dir").await?;

pub async fn list(&self, path: &str) -> Result<Vec<Entry>, Error>

List entries within a given directory.

Notes

Listing recursively

This function only read the children of the given directory. To read all entries recursively, use Operator::list_with("path").delimiter("") instead.

Streaming

This function will read all entries in the given directory. It could take very long time and consume a lot of memory if the directory contains a lot of entries.

In order to avoid this, you can use Operator::lister to list entries in a streaming way.

Metadata

The only metadata that is guaranteed to be available is the Mode. For fetching more metadata, please use Operator::list_with and metakey.

Examples

use opendal::EntryMode;
use opendal::Metakey;
use opendal::Operator;
let mut entries = op.list("path/to/dir/").await?;
for entry in entries {
    match entry.metadata().mode() {
        EntryMode::FILE => {
            println!("Handling file")
        }
        EntryMode::DIR => {
            println!("Handling dir {}", entry.path())
        }
        EntryMode::Unknown => continue,
    }
}

pub fn list_with(&self, path: &str) -> FutureList

List entries within a given directory with options.

Notes

For streaming

This function will read all entries in the given directory. It could take very long time and consume a lot of memory if the directory contains a lot of entries.

In order to avoid this, you can use Operator::lister to list entries in a streaming way.

Metadata

The only metadata that is guaranteed to be available is the Mode. For fetching more metadata, please specify the metakey.

Examples

List entries with prefix

This function can also be used to list entries in recursive way.

use opendal::EntryMode;
use opendal::Metakey;
use opendal::Operator;
let mut entries = op.list_with("prefix/").delimiter("").await?;
for entry in entries {
    match entry.metadata().mode() {
        EntryMode::FILE => {
            println!("Handling file")
        }
        EntryMode::DIR => {
            println!("Handling dir like start a new list via meta.path()")
        }
        EntryMode::Unknown => continue,
    }
}

List entries with metakey for more metadata

use opendal::EntryMode;
use opendal::Metakey;
use opendal::Operator;
let mut entries = op
    .list_with("dir/")
    .metakey(Metakey::ContentLength | Metakey::LastModified)
    .await?;
for entry in entries {
    let meta = entry.metadata();
    match meta.mode() {
        EntryMode::FILE => {
            println!(
                "Handling file {} with size {}",
                entry.path(),
                meta.content_length()
            )
        }
        EntryMode::DIR => {
            println!("Handling dir {}", entry.path())
        }
        EntryMode::Unknown => continue,
    }
}

pub async fn lister(&self, path: &str) -> Result<Lister, Error>

List entries within a given directory as a stream.

This function will create a new handle to list entries.

An error will be returned if given path doesn’t end with /.

Notes

Listing recursively

This function only read the children of the given directory. To read all entries recursively, use Operator::lister_with and delimiter("") instead.

Metadata

The only metadata that is guaranteed to be available is the Mode. For fetching more metadata, please use Operator::lister_with and metakey.

Examples

use futures::TryStreamExt;
use opendal::EntryMode;
use opendal::Metakey;
use opendal::Operator;
let mut ds = op.lister("path/to/dir/").await?;
while let Some(mut de) = ds.try_next().await? {
    match de.metadata().mode() {
        EntryMode::FILE => {
            println!("Handling file")
        }
        EntryMode::DIR => {
            println!("Handling dir like start a new list via meta.path()")
        }
        EntryMode::Unknown => continue,
    }
}

pub fn lister_with(&self, path: &str) -> FutureLister

List entries within a given directory as a stream with options.

This function will create a new handle to list entries.

An error will be returned if given path doesn’t end with /.

Examples

List current dir

use futures::TryStreamExt;
use opendal::EntryMode;
use opendal::Metakey;
use opendal::Operator;
let mut ds = op
    .lister_with("path/to/dir/")
    .limit(10)
    .start_after("start")
    .await?;
while let Some(mut entry) = ds.try_next().await? {
    match entry.metadata().mode() {
        EntryMode::FILE => {
            println!("Handling file {}", entry.path())
        }
        EntryMode::DIR => {
            println!("Handling dir {}", entry.path())
        }
        EntryMode::Unknown => continue,
    }
}

List all files recursively

use futures::TryStreamExt;
use opendal::EntryMode;
use opendal::Metakey;
use opendal::Operator;
let mut ds = op.lister_with("path/to/dir/").delimiter("").await?;
while let Some(mut entry) = ds.try_next().await? {
    match entry.metadata().mode() {
        EntryMode::FILE => {
            println!("Handling file {}", entry.path())
        }
        EntryMode::DIR => {
            println!("Handling dir {}", entry.path())
        }
        EntryMode::Unknown => continue,
    }
}

List files with required metadata

use futures::TryStreamExt;
use opendal::EntryMode;
use opendal::Metakey;
use opendal::Operator;
let mut ds = op
    .lister_with("path/to/dir/")
    .metakey(Metakey::ContentLength | Metakey::LastModified)
    .await?;
while let Some(mut entry) = ds.try_next().await? {
    let meta = entry.metadata();
    match meta.mode() {
        EntryMode::FILE => {
            println!(
                "Handling file {} with size {}",
                entry.path(),
                meta.content_length()
            )
        }
        EntryMode::DIR => {
            println!("Handling dir {}", entry.path())
        }
        EntryMode::Unknown => continue,
    }
}

impl Operator

Operator presign API.

pub async fn presign_stat( &self, path: &str, expire: Duration ) -> Result<PresignedRequest, Error>

Presign an operation for stat(head).

Example

use anyhow::Result;
use futures::io;
use opendal::Operator;
use std::time::Duration;

#[tokio::main]
async fn test(op: Operator) -> Result<()> {
    let signed_req = op.presign_stat("test",Duration::from_secs(3600)).await?;
    let req = http::Request::builder()
        .method(signed_req.method())
        .uri(signed_req.uri())
        .body(())?;

pub async fn presign_read( &self, path: &str, expire: Duration ) -> Result<PresignedRequest, Error>

Presign an operation for read.

Example

use anyhow::Result;
use futures::io;
use opendal::Operator;
use std::time::Duration;

#[tokio::main]
async fn test(op: Operator) -> Result<()> {
    let signed_req = op.presign_read("test.txt", Duration::from_secs(3600)).await?;

signed_req.method(): GET

signed_req.uri(): https://s3.amazonaws.com/examplebucket/test.txt?X-Amz-Algorithm=AWS4-HMAC-SHA256&X-Amz-Credential=access_key_id/20130721/us-east-1/s3/aws4_request&X-Amz-Date=20130721T201207Z&X-Amz-Expires=86400&X-Amz-SignedHeaders=host&X-Amz-Signature=<signature-value>

signed_req.headers(): { "host": "s3.amazonaws.com" }

We can download this file via curl or other tools without credentials:

curl "https://s3.amazonaws.com/examplebucket/test.txt?X-Amz-Algorithm=AWS4-HMAC-SHA256&X-Amz-Credential=access_key_id/20130721/us-east-1/s3/aws4_request&X-Amz-Date=20130721T201207Z&X-Amz-Expires=86400&X-Amz-SignedHeaders=host&X-Amz-Signature=<signature-value>" -O /tmp/test.txt

pub fn presign_read_with( &self, path: &str, expire: Duration ) -> FuturePresignRead

Presign an operation for read option described in OpenDAL [RFC-1735][crate::docs::rfcs::rfc_1735_operation_extension].

You can pass OpRead to this method to specify the content disposition.

Example

use anyhow::Result;
use futures::io;
use opendal::Operator;
use std::time::Duration;

#[tokio::main]
async fn test(op: Operator) -> Result<()> {
    let signed_req = op
        .presign_read_with("test.txt", Duration::from_secs(3600))
        .override_content_disposition("attachment; filename=\"othertext.txt\"")
        .await?;

pub async fn presign_write( &self, path: &str, expire: Duration ) -> Result<PresignedRequest, Error>

Presign an operation for write.

Example

use anyhow::Result;
use futures::io;
use opendal::Operator;
use std::time::Duration;

#[tokio::main]
async fn test(op: Operator) -> Result<()> {
    let signed_req = op.presign_write("test.txt", Duration::from_secs(3600)).await?;

signed_req.method(): PUT

signed_req.uri(): https://s3.amazonaws.com/examplebucket/test.txt?X-Amz-Algorithm=AWS4-HMAC-SHA256&X-Amz-Credential=access_key_id/20130721/us-east-1/s3/aws4_request&X-Amz-Date=20130721T201207Z&X-Amz-Expires=86400&X-Amz-SignedHeaders=host&X-Amz-Signature=<signature-value>

signed_req.headers(): { "host": "s3.amazonaws.com" }

We can upload file as this file via curl or other tools without credential:

curl -X PUT "https://s3.amazonaws.com/examplebucket/test.txt?X-Amz-Algorithm=AWS4-HMAC-SHA256&X-Amz-Credential=access_key_id/20130721/us-east-1/s3/aws4_request&X-Amz-Date=20130721T201207Z&X-Amz-Expires=86400&X-Amz-SignedHeaders=host&X-Amz-Signature=<signature-value>" -d "Hello, World!"

pub fn presign_write_with( &self, path: &str, expire: Duration ) -> FuturePresignWrite

Presign an operation for write with option described in OpenDAL [RFC-0661][crate::docs::rfcs::rfc_0661_path_in_accessor]

You can pass OpWrite to this method to specify the content length and content type.

Example

use anyhow::Result;
use futures::io;
use opendal::Operator;
use std::time::Duration;

#[tokio::main]
async fn test(op: Operator) -> Result<()> {
    let signed_req = op.presign_write_with("test", Duration::from_secs(3600))
                       .content_type("text/csv").await?;
    let req = http::Request::builder()
        .method(signed_req.method())
        .uri(signed_req.uri())
        .body(())?;

impl Operator

Operator build API

Operator should be built via [OperatorBuilder]. We recommend to use Operator::new to get started:

use opendal::services::Fs;
use opendal::Operator;
#[tokio::main]
async fn main() -> Result<()> {
    // Create fs backend builder.
    let mut builder = Fs::default();
    // Set the root for fs, all operations will happen under this root.
    //
    // NOTE: the root must be absolute path.
    builder.root("/tmp");

    // Build an `Operator` to start operating the storage.
    let op: Operator = Operator::new(builder)?.finish();

    Ok(())
}

pub fn new<B>(ab: B) -> Result<OperatorBuilder<impl Accessor>, Error>
where B: Builder,

Create a new operator with input builder.

OpenDAL will call builder.build() internally, so we don’t need to import opendal::Builder trait.

Examples

Read more backend init examples in examples.

use opendal::services::Fs;
use opendal::Operator;
#[tokio::main]
async fn main() -> Result<()> {
    // Create fs backend builder.
    let mut builder = Fs::default();
    // Set the root for fs, all operations will happen under this root.
    //
    // NOTE: the root must be absolute path.
    builder.root("/tmp");

    // Build an `Operator` to start operating the storage.
    let op: Operator = Operator::new(builder)?.finish();

    Ok(())
}

pub fn from_map<B>( map: HashMap<String, String> ) -> Result<OperatorBuilder<impl Accessor>, Error>
where B: Builder,

Create a new operator from given map.

Notes

from_map is using static dispatch layers which is zero cost. via_map is using dynamic dispatch layers which has a bit runtime overhead with an extra vtable lookup and unable to inline. But from_map requires generic type parameter which is not always easy to be used.

Examples

use std::collections::HashMap;

use opendal::services::Fs;
use opendal::Operator;
#[tokio::main]
async fn main() -> Result<()> {
    let map = HashMap::from([
        // Set the root for fs, all operations will happen under this root.
        //
        // NOTE: the root must be absolute path.
        ("root".to_string(), "/tmp".to_string()),
    ]);

    // Build an `Operator` to start operating the storage.
    let op: Operator = Operator::from_map::<Fs>(map)?.finish();

    Ok(())
}

pub fn via_map( scheme: Scheme, map: HashMap<String, String> ) -> Result<Operator, Error>

Create a new operator from given scheme and map.

Notes

from_map is using static dispatch layers which is zero cost. via_map is using dynamic dispatch layers which has a bit runtime overhead with an extra vtable lookup and unable to inline. But from_map requires generic type parameter which is not always easy to be used.

Examples

use std::collections::HashMap;

use opendal::Operator;
use opendal::Scheme;
#[tokio::main]
async fn main() -> Result<()> {
    let map = HashMap::from([
        // Set the root for fs, all operations will happen under this root.
        //
        // NOTE: the root must be absolute path.
        ("root".to_string(), "/tmp".to_string()),
    ]);

    // Build an `Operator` to start operating the storage.
    let op: Operator = Operator::via_map(Scheme::Fs, map)?;

    Ok(())
}

pub fn layer<L>(self, layer: L) -> Operator
where L: Layer<Arc<dyn Accessor<BlockingPager = Box<dyn BlockingPage>, BlockingWriter = Box<dyn BlockingWrite>, Writer = Box<dyn Write>, Pager = Box<dyn Page>, Reader = Box<dyn Read>, BlockingReader = Box<dyn BlockingRead>>>>,

Create a new layer with dynamic dispatch.

Notes

OperatorBuilder::layer() is using static dispatch which is zero cost. Operator::layer() is using dynamic dispatch which has a bit runtime overhead with an extra vtable lookup and unable to inline.

It’s always recommended to use OperatorBuilder::layer() instead.

Examples

use opendal::layers::LoggingLayer;
use opendal::services::Fs;
use opendal::Operator;

let op = Operator::new(Fs::default())?.finish();
let op = op.layer(LoggingLayer::default());
// All operations will go through the new_layer
let _ = op.read("test_file").await?;