Module object_store::util

Functions

collect

Collect all entries from the Lister.

join_dir

Join two paths and normalize the output dir.

join_path

Push child to parent dir and normalize the output path.

normalize_dir

Modified from the opendal::raw::normalize_root

normalize_path

Make sure all operation are constructed by normalized path:

with_instrument_layers

Attaches instrument layers to the object store.